Pre-release Improvements (#20)
* adjustments

* Multiple adjustments, based on ongoing testing

* Apply CS fixes

* Various adjustments

* Improve `ColumnComparator` to validate relation vs scalar conflicts, enhance exception messages, and simplify test fixtures

* Apply CS fixes

* Many-to-many pivot interface updated

* Apply CS fixes

* Entity manager refactoring

* Apply CS fixes

* Resolving deptrac dependencies

* Indexes validation updated

* Column comparator fix

* circular foreign key logic updated

* another foreign key logic fix

* migration generation logic flaw fixed

* DETACHED state added, pgsql migration bug fixed

* Apply CS fixes

* Single quotes in migrations to not force escaping for pgsql

* Diff comparison normalisation

* Diff comparison normalisation

* Apply CS fixes

* mysql algorithm inplace fix

---------

+++ tests/Modules/Database/TypeMapperTest.php
@@ -14,7 +14,7 @@ class TypeMapperTest extends TestCase {
         $mapper = new MySqlTypeMapper();
 
         $this->assertSame('INT', $mapper->getDatabaseType('int'));
-        $this->assertSame('FLOAT', $mapper->getDatabaseType('float'));
+        $this->assertSame('DOUBLE', $mapper->getDatabaseType('float'));
         $this->assertSame('VARCHAR(255)', $mapper->getDatabaseType('string'));
         $this->assertSame('TINYINT(1)', $mapper->getDatabaseType('bool'));
         $this->assertSame('TEXT', $mapper->getDatabaseType('mixed'));
